+++ src/Extension.php
@@ -18,11 +18,31 @@ class Extension extends AbstractExtension
     public function getFunctions(): array
     {
         return [
-            new TwigFunction('umanit_image_figure', [Runtime::class, 'getUmanitImageFigure']),
-            new TwigFunction('umanit_image_figure_lazy_load', [Runtime::class, 'getUmanitImageFigureLazyLoad']),
-            new TwigFunction('umanit_image_picture', [Runtime::class, 'getUmanitImagePicture']),
-            new TwigFunction('umanit_image_picture_lazy_load', [Runtime::class, 'getUmanitImagePictureLazyLoad']),
-            new TwigFunction('umanit_image_srcset', [Runtime::class, 'getUmanitImageSrcset']),
+            new TwigFunction(
+                'umanit_image_figure',
+                [Runtime::class, 'getUmanitImageFigure'],
+                ['is_safe' => ['html']]
+            ),
+            new TwigFunction(
+                'umanit_image_figure_lazy_load',
+                [Runtime::class, 'getUmanitImageFigureLazyLoad'],
+                ['is_safe' => ['html']]
+            ),
+            new TwigFunction(
+                'umanit_image_picture',
+                [Runtime::class, 'getUmanitImagePicture'],
+                ['is_safe' => ['html']]
+            ),
+            new TwigFunction(
+                'umanit_image_picture_lazy_load',
+                [Runtime::class, 'getUmanitImagePictureLazyLoad'],
+                ['is_safe' => ['html']]
+            ),
+            new TwigFunction(
+                'umanit_image_srcset',
+                [Runtime::class, 'getUmanitImageSrcset'],
+                ['is_safe' => ['html']]
+            ),
         ];
     }
 }